Rework filter test case organization
Reorganize the tests such that each individual pattern match comparison
has its own TestCase instance.  This means that the tests passed/failed
count accurately reflects the number of tests done, and that all of the
failures are displayed individually in the test runner output (instead
of merely the first failure).